requirements.txt
python-telegram-bot==13.7
pyrogram==2.0.49
requests==2.25.1
openai==0.27.0
flask==2.1.0